Use Text for observations
The benefit of this is that Text has a much faster Hashable instance,
which allows consumers of the Hoed trace to be much more efficient.
In absence of pointer equality, a fast hashing operation is paramount
when dealing with computation graphs of >1 million nodes.

I would not expect this change to be an improvement in terms of space usage
in the observation trace, because the strings in the observation trace
(including Con events) are usually shared.